+++ SpiritEngine/src/SpiritEngine/Data/Models/Items/Equipment/Weapon/Weapon.java
@@ -11,12 +11,12 @@ public class Weapon extends Item {
     public Damage dmgS;
     public Damage dmgM;
     public Critical critical;
-    public double range; //In ft
+    public int range; //In ft
     public double weight; //In lbs
     public Type[] type;
     public Special[] special;
 
-    public Weapon(String name, int id, int maxCapacity, double cost, Damage dmgS, Damage dmgM, Critical critical, double range, double weight, Type[] type, Special[] special) {
+    public Weapon(String name, int id, int maxCapacity, double cost, Damage dmgS, Damage dmgM, Critical critical, int range, double weight, Type[] type, Special[] special) {
         super(name, id, cost);
         this.name = name;
         this.id = id;
